+++ src/sentry/search/eap/span_columns.py
@@ -15,6 +15,7 @@ from sentry.search.eap.columns import (
     simple_measurements_field,
     simple_sentry_field, project_term_resolver,
 )
+from sentry.search.eap.common_columns import COMMON_COLUMNS
 from sentry.search.events.constants import SPAN_MODULE_CATEGORY_VALUES
 from sentry.search.events.types import SnubaParams
 from sentry.search.utils import DEVICE_CLASS
@@ -22,7 +23,8 @@ from sentry.utils.validators import is_event_id, is_span_id
 
 SPAN_ATTRIBUTE_DEFINITIONS = {
     column.public_alias: column
-    for column in [
+    for column in COMMON_COLUMNS
+    + [
         ResolvedColumn(
             public_alias="id",
             internal_name="sentry.span_id",
@@ -34,24 +36,6 @@ SPAN_ATTRIBUTE_DEFINITIONS = {
             internal_name="sentry.parent_span_id",
             search_type="string",
             validator=is_span_id,
-        ),
-        ResolvedColumn(
-            public_alias="organization.id",
-            internal_name="sentry.organization_id",
-            search_type="string",
-        ),
-        ResolvedColumn(
-            public_alias="project.id",
-            internal_name="sentry.project_id",
-            internal_type=constants.INT,
-            search_type="string",
-        ),
-        ResolvedColumn(
-            public_alias="project_id",
-            internal_name="sentry.project_id",
-            internal_type=constants.INT,
-            search_type="string",
-            secondary_alias=True,
         ),
         ResolvedColumn(
             public_alias="span.action",
